backend/requirements.txt
sentence-transformers
numpy
flask
PyPDF2
python-docx
